swagger added

+++ app/Http/Controllers/AppController.php
@@ -5,16 +5,69 @@ namespace App\Http\Controllers;
 use App\App;
 use App\Http\Resources\AppResource;
 use Illuminate\Http\Request;
+use http\Env\Response;
 
 class AppController extends Controller
 {
     /**
+     *
+     * @param Request $request
+     * @return Response|string
+     */
+
+
+    /**
      * Display a listing of the resource.
      *
-     * @return \Illuminate\Http\Response
+     * @return Response
+     */
+
+    /**
+     * @SWG\Get(
+     *     path="api/apps",
+     *     description="Return all apps",
+     *     @SWG\Parameter(
+     *         name="name",
+     *         in="query",
+     *         type="string",
+     *         description="app name",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="description",
+     *         in="query",
+     *         type="string",
+     *         description="app description",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="icon",
+     *         in="query",
+     *         type="string",
+     *         description="app icon",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="creator",
+     *         in="query",
+     *         type="string",
+     *         description="app creator",
+     *         required=true,
+     *     ),
+     *     @SWG\Response(
+     *         response=200,
+     *         description="Successful operation",
+     *     ),
+     *     @SWG\Response(
+     *          response=403,
+     *          description="Forbidden"
+     *      )
+     *
+     * )
      */
     public function index()
     {
+        abort_if(Gate::denies('project_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return AppResource::collection(App::get());
     }
 
@@ -22,6 +75,53 @@ class AppController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     */
+
+    /**
+     * @SWG\Post(
+     *     path="api/apps",
+     *     description="add new record to app table",
+     *     @SWG\Parameter(
+     *         name="name",
+     *         in="query",
+     *         type="string",
+     *         description="app name",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="description",
+     *         in="query",
+     *         type="string",
+     *         description="app description",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="icon",
+     *         in="query",
+     *         type="string",
+     *         description="app icon",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="creator",
+     *         in="query",
+     *         type="string",
+     *         description="app creator",
+     *         required=true,
+     *     ),
+     *     @SWG\Response(
+     *         response=201,
+     *         description="OK",
+     *     ),
+     *     @SWG\Response(
+     *         response=400,
+     *         description="Bad request",
+     *     ),
+     *     @SWG\Response(
+     *         response=403,
+     *         description="Forbidden",
+     *     )
+     * )
      */
 
     public function store(Request $request)
@@ -32,20 +132,75 @@ class AppController extends Controller
             'icon' => 'required|url',
             'creator' => 'required|max:50'
         ]);
+//        $appData = $request->all();
+
+//        if (empty($appData['name']) && empty($appData['description']) && empty($appData['icon']) && empty($appData['creator'])) {
+//            return new \Exception('Missing data', 404);
+//        }
 
         $app = App::create($request->all());
 
-        return new AppResource($app);
+        return (new AppResource($app))
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  int  $id
+     * @param  int  $app
      * @return \Illuminate\Http\Response
+     */
+
+    /**
+     * @SWG\Get(
+     *     path="api/apps/{app}",
+     *     description="show one record in app table",
+     *     @SWG\Parameter(
+     *         name="name",
+     *         in="query",
+     *         type="string",
+     *         description="app name",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="description",
+     *         in="query",
+     *         type="string",
+     *         description="app description",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="icon",
+     *         in="query",
+     *         type="string",
+     *         description="app icon",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="creator",
+     *         in="query",
+     *         type="string",
+     *         description="app creator",
+     *         required=true,
+     *     ),
+     *     @SWG\Response(
+     *         response=200,
+     *         description="OK",
+     *     ),
+     *     @SWG\Response(
+     *         response=400,
+     *         description="Bad request",
+     *     ),
+     *     @SWG\Response(
+     *         response=403,
+     *         description="Forbidden",
+     *     )
+     * )
      */
     public function show(App $app)
     {
+        abort_if(Gate::denies('project_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
         return new AppResource($app);
 
     }
@@ -55,8 +210,59 @@ class AppController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
+     * @param  int  $app
      * @return \Illuminate\Http\Response
+     */
+
+    /**
+     * @SWG\Put(
+     *     path="api/apps/{app}",
+     *     description="update record in app table",
+     *     @SWG\Parameter(
+     *         name="name",
+     *         in="query",
+     *         type="string",
+     *         description="app name",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="description",
+     *         in="query",
+     *         type="string",
+     *         description="app description",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="icon",
+     *         in="query",
+     *         type="string",
+     *         description="app icon",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="creator",
+     *         in="query",
+     *         type="string",
+     *         description="app creator",
+     *         required=true,
+     *     ),
+     *     @SWG\Response(
+     *         response=202,
+     *         description="OK",
+     *     ),
+     *     @SWG\Response(
+     *         response=400,
+     *         description="Bad request",
+     *     ),
+     *     @SWG\Response(
+     *         response=403,
+     *         description="Forbidden",
+     *     ),
+     *     @SWG\Response(
+     *         response=404,
+     *         description="Not found",
+     *     )
+     * )
      */
     public function update(Request $request, App $app)
     {
@@ -69,19 +275,70 @@ class AppController extends Controller
 
         $app->update($request->all());
 
-        return new AppResource($app);
+        return (new AppResource($app))
+            ->response()
+            ->setStatusCode(Response::HTTP_CREATED);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  int  $id
+     * @param  int  $app
      * @return \Illuminate\Http\Response
+     */
+
+    /**
+     * @SWG\Delete(
+     *     path="api/apps/{app}",
+     *     description="delete record from app table",
+     *     @SWG\Parameter(
+     *         name="name",
+     *         in="query",
+     *         type="string",
+     *         description="app name",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="description",
+     *         in="query",
+     *         type="string",
+     *         description="app description",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="icon",
+     *         in="query",
+     *         type="string",
+     *         description="app icon",
+     *         required=true,
+     *     ),
+     *     @SWG\Parameter(
+     *         name="creator",
+     *         in="query",
+     *         type="string",
+     *         description="app creator",
+     *         required=true,
+     *     ),
+     *     @SWG\Response(
+     *         response=204,
+     *         description="OK",
+     *     ),
+     *     @SWG\Response(
+     *         response=403,
+     *         description="Forbidden",
+     *     ),
+     *     @SWG\Response(
+     *         response=404,
+     *         description="Not found",
+     *     )
+     * )
      */
     public function destroy(App $app)
     {
+        abort_if(Gate::denies('project_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+
         $app->delete();
 
-        return response('massege' , 'deleted');
+        return response(null, Response::HTTP_NO_CONTENT);
     }
 }
